requirements.txt
torch>=1.10.0
torchvision>=0.11.1
opencv-python>=4.5.4
numpy>=1.21.2
scikit-learn>=1.0
matplotlib>=3.5.0
pyyaml>=6.0
tqdm>=4.62.3